fix(loop): update pMostBoWMatchesKF + float Sim3 thresholds + fr1_room cfg
- LoopClosing.cc (point DetectCommonRegionsFromBoW): pMostBoWMatchesKF was never
  updated to the covisible KF with most BoW matches (the assignment was commented
  out), so Sim3Solver looked up correspondences in the wrong keyframe. The
  line-version already had this fix. Also add stage/bestInl diagnostics to the
  line BoW detection.
- Sim3Solver.h: mvnMaxError/mvSigmaSquare were std::vector<size_t>, truncating
  float reprojection thresholds; changed to std::vector<float>.
- Add TUM freiburg1_room SLAM + gaussian configs (copied from freiburg1_desk,
  same camera) and a diag gaussian variant.

Verified on fr1_room (baseline A): still no accepted loop; monocular ATE is
0.89 m (scale 0.795) so drift is the blocker there. A couple candidates do pass
Sim3 (stage 4-5) but fail the 3-KF covisibility consistency. office0 RGB-D is
4.8 mm ATE and correctly has no loop to close.

+++ ORB-SLAM3/include/Sim3Solver.h
@@ -75,10 +75,10 @@ protected:
     std::vector<MapPoint*> mvpMapPoints2;
     std::vector<MapPoint*> mvpMatches12;
     std::vector<size_t> mvnIndices1;
-    std::vector<size_t> mvSigmaSquare1;
-    std::vector<size_t> mvSigmaSquare2;
-    std::vector<size_t> mvnMaxError1;
-    std::vector<size_t> mvnMaxError2;
+    std::vector<float> mvSigmaSquare1;
+    std::vector<float> mvSigmaSquare2;
+    std::vector<float> mvnMaxError1;
+    std::vector<float> mvnMaxError2;
 
     int N;
     int mN1;
